haveNewOrdersMArket servise added and market and customer folders deleted
