--- server/src/api/customer/dto/create-customer.dto.ts
@@ -1,29 +1,0 @@
-import {
-  IsNotEmpty,
-  IsOptional,
-  IsPhoneNumber,
-  IsString,
-  IsUUID,
-} from 'class-validator';
-
-export class CreateCustomerDto {
-  @IsNotEmpty()
-  @IsUUID()
-  market_id: string;
-
-  @IsNotEmpty()
-  @IsString()
-  client_name: string;
-
-  @IsNotEmpty()
-  @IsPhoneNumber('UZ')
-  client_phone_number: string;
-
-  @IsNotEmpty()
-  @IsUUID()
-  district_id: string;
-
-  @IsOptional()
-  @IsString()
-  address?: string;
-}
