--- server/src/api/customer/dto/update-customer.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateCustomerDto } from './create-customer.dto';
-
-export class UpdateCustomerDto extends PartialType(CreateCustomerDto) {}
