--- server/src/api/market/dto/login-market.dto.ts
@@ -1,11 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class LoginMarketDto {
-  @IsString()
-  @IsNotEmpty()
-  phone_number: string;
-
-  @IsString()
-  @IsNotEmpty()
-  password: string;
-}
