--- server/src/api/market/dto/update-market.dto.ts
@@ -1,34 +1,0 @@
-import { IsEnum, IsNumber, IsOptional, IsString, Min } from 'class-validator';
-import { AddOrder, Status } from 'src/common/enums';
-
-export class UpdateMarketDto {
-  @IsOptional()
-  @IsString()
-  market_name?: string;
-
-  @IsOptional()
-  @IsString()
-  phone_number?: string;
-
-  @IsOptional()
-  @IsString()
-  password?: string;
-
-  @IsOptional()
-  @IsEnum(Status)
-  status?: Status;
-
-  @IsOptional()
-  @IsEnum(AddOrder)
-  add_order?: AddOrder;
-
-  @IsOptional()
-  @IsNumber()
-  @Min(0)
-  tariff_home?: number;
-
-  @IsOptional()
-  @IsNumber()
-  @Min(0)
-  tariff_center?: number;
-}
