--- server/src/api/market/dto/update-own.dto.ts
@@ -1,11 +1,0 @@
-import { IsOptional, IsPhoneNumber, IsString } from 'class-validator';
-
-export class UpdateOwnMarketDto {
-  @IsOptional()
-  @IsPhoneNumber('UZ')
-  phone_number?: string;
-
-  @IsOptional()
-  @IsString()
-  password?: string;
-}
