+++ server/src/api/order/order.module.ts
@@ -7,7 +7,6 @@ import { OrderItemEntity } from 'src/core/entity/order-item.entity';
 import { ProductEntity } from 'src/core/entity/product.entity';
 import { MarketEntity } from 'src/core/entity/market.entity';
 import { ProductService } from '../product/product.service';
-import { MarketService } from '../market/market.service';
 import { CashBoxModule } from '../cash-box/cash-box.module';
 import { BcryptEncryption } from 'src/infrastructure/lib/bcrypt';
 import { Token } from 'src/infrastructure/lib/token-generator/token';
@@ -31,12 +30,6 @@ import { UserEntity } from 'src/core/entity/users.entity';
     CashBoxModule,
   ],
   controllers: [OrderController],
-  providers: [
-    OrderService,
-    ProductService,
-    MarketService,
-    BcryptEncryption,
-    Token,
-  ],
+  providers: [OrderService, ProductService, BcryptEncryption, Token],
 })
 export class OrderModule {}
