+++ server/src/api/order/order.service.ts
@@ -150,6 +150,38 @@ export class OrderService extends BaseService<CreateOrderDto, OrderEntity> {
     }
   }
 
+  async haveNewOrderMarkets() {
+    try {
+      const allNewOrders = await this.orderRepo.find({
+        where: { status: Order_status.NEW },
+      });
+
+      if (!allNewOrders.length) {
+        return successRes([], 200, 'No new orders');
+      }
+
+      const uniqueMarketIds = Array.from(
+        new Set(allNewOrders.map((order) => order.market_id)),
+      );
+
+      const allUniqueMarkets = await this.userRepo.find({
+        where: { id: In(uniqueMarketIds), role: Roles.MARKET },
+      });
+
+      return successRes(
+        {
+          count: allUniqueMarkets.length,
+          markets: allUniqueMarkets,
+        },
+        200,
+        'Markets with new orders',
+      );
+    } catch (error) {
+      console.error(error);
+      throw error;
+    }
+  }
+
   async newOrdersByMarketId(id: string) {
     try {
       const market = await this.marketRepo.findOne({ where: { id } });
